"Refactor data service and screens to update accounts receivable creation and hauling advice list"

+++ lib/data_service.dart
@@ -143,7 +143,7 @@ class DataService {
 
   Future<void> createAccountsReceivable({
     required String billingNo,
-    required int totalAmount,
+    required double totalAmount,
     required String billingDate,
     required int amountPaid,
     String? paymentDate,
